refactor(expo): break the upload import cycles
`ui-components/index.ts` exported UploadProgressBar, which imported
UploadStage out of `screens/`, which imported FileThumbnail back through
the ui-components barrel. Separately, ClientProfilePhotoUploader pulled
useSnackbar from the `hooks` barrel, which reaches into `screens/`, which
import ui-components back. Both resolve at render time today, but a cycle
can hand a module `undefined` at evaluation time.

Move UploadStage into ui-components beside the bar that opens it (a screen
was the wrong home for something the global bar owns), leaving the
docs-domain useDocsUpload hook in screens/Client/Docs. Both now import
their collaborators by path rather than through a barrel.

Verified with madge: the branch adds no cycles relative to main, down from
one. The other 32 are pre-existing and untouched.
