+++ libs/expo/betterangels/src/lib/screens/Client/Docs/UploadModal/index.tsx
@@ -10,7 +10,7 @@ import { useEffect, useState } from 'react';
 import { ScrollView, View } from 'react-native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { ClientDocumentNamespaceEnum } from '../../../../apollo';
-import { useDocsUpload } from '../UploadStage/useDocsUpload';
+import { useDocsUpload } from '../useDocsUpload';
 import FileUploadTab from './FileUploadTab';
 import { DocUploads, IUploadModalProps } from './types';
